+++ Carbon.IIS/Functions/Get-CIisApplication.ps1
@@ -14,8 +14,6 @@ function Get-CIisApplication
      * `CommitChanges()` - Persists any configuration changes made to the object back into IIS's configuration files.
      * `PhysicalPath { get; }` - The physical path to the application.
 
-    Beginning with Carbon 2.0.1, this function is available only if IIS is installed.
-
     .OUTPUTS
     Microsoft.Web.Administration.Application.
 
@@ -41,8 +39,8 @@ function Get-CIisApplication
         [Parameter(Mandatory)]
         [String] $SiteName,
 
-        # The name of the application.  Default is to return all applications running under the website `$SiteName`.
-        [Alias('Name')]
+        # The path/name of the application. Default is to return all applications running under the website given by
+        # the `SiteName` parameter.
         [String] $VirtualPath
     )
 
@@ -55,11 +53,13 @@ function Get-CIisApplication
         return
     }
 
+    $VirtualPath = $VirtualPath | ConvertTo-CIisVirtualPath
+
     $site.Applications |
         Where-Object {
-            if( $VirtualPath )
+            if( $PSBoundParameters.ContainsKey('VirtualPath') )
             {
-                return ($_.Path -eq "/$VirtualPath")
+                return ($_.Path -eq $VirtualPath)
             }
             return $true
         } |