+++ CHANGELOG.md
@@ -69,6 +69,16 @@ Removed `SiteName` parameter aliases. Update usages of the `SiteName` parameter 
 * Get-CIisWebsite
 * Install-CIisWebsite
 
+Removed the `Path` parameter alias on the `Install-CIisApplication` function. Update usages to `PhysicalPath`.
+
+Renamed the `VirtualPath` parameter to `Path` on the `Get-CIisApplication` and `Install-CIisApplication` functions.
+Update usages.
+
+Removed the `Name` paramater aliases. Update usages of the `Name` parameter on these functions to `Path`:
+* Get-CIisApplication
+* Install-CIisApplication
+
+
 ## Added
 
 * Carbon.IIS now supports
@@ -86,6 +96,7 @@ Pass the application name to the `ApplicationPath` parameter.
 * `Get-CIisWebsite` now writes an error if a specific website doesn't exist.
 * The `Join-CIisVirtualPath` function's `ChildPath` parameter is now required. Usages that don't have a `ChildPath`
 argument should switch to `ConvertTo-CIisVirtualPath`.
+* Renamed `VirtualPath` parameter on `Get-CIisApplication` and `Install-CIisApplication` to `Path`.
 
 ## Fixed
 
@@ -101,3 +112,28 @@ and `Remove-CIisWebsite` (for `Uninstall-CIisWebsite`).
 * The objects returned from the `Get-CIisHttpRedirect` function changed to
 `Microsoft.Web.Administration.ConfigurationSection` objects. They no longer have `ChildOnly`, `Destination`, `Enabled`, `ExactDestination`, and `HttpResponseStatus`
 properties. Use `GetAttributeValue` or `SetAttributeValue` to get/set values instead.
+* `Path` parameter alias on functions
+    * Disable-CIisSecurityAuthentication
+    * Enable-CIisSecurityAuthentication
+    * Enable-CIisSecurityAuthentication
+    * Enable-CIisSsl
+    * Get-CIisApplication
+    * Get-CIisConfigurationSection
+    * Get-CIisHttpHeader
+    * Get-CIisHttpRedirect
+    * Get-CIisMimeMap
+    * Get-CIisSecurityAuthentication
+    * Install-CIisApplication
+    * Set-CIisHttpHeader
+    * Set-CIisHttpRedirect
+    * Set-CIisWindowsAuthentication
+    * Test-CIisConfigurationSection
+    * Test-CIisSecurityAuthentication
+* `SiteName` parameter alias on functions:
+    * Get-CIisWebsite
+    * Install-CIisWebsite
+* `Path` parameter alias to `PhysicalPath` on function `Install-CIisApplication`.
+* `Name` parameter alias on functions:
+    * Get-CIisApplication
+    * Install-CIisApplication
+